updated the file download elements
